Move nearby request handling to request controller
